Add pyPasswordGenerator.py for creating a random password
This commit adds a new file, pyPasswordGenerator.py, which serves as a simple password generator. The file allows the user from the standard input to specify the number of letters, symbols, and numbers they want in their password. The program then creates a password according to the user's specifications and prints it. The order of the characters in the generated password is randomized for additional security.